refactor: web module extraction - GlobalExceptionHandler migration (#387, #388)
- Extract GlobalExceptionHandler to module-web for web layer isolation
- Remove PiiMaskingFilter (moved to module-infra security layer)
- Update Spring isolation tests for new module structure
- Add verify-guardrails skill to project verification suite
- Update ArchTest with proper assertions for documentation-only tests
