+++ module-app/src/main/java/maple/expectation/monitoring/ai/AiSreService.java
@@ -9,9 +9,9 @@ import java.util.concurrent.Executor;
 import lombok.extern.slf4j.Slf4j;
 import maple.expectation.infrastructure.executor.LogicExecutor;
 import maple.expectation.infrastructure.executor.TaskContext;
+import maple.expectation.infrastructure.monitoring.security.PiiMaskingFilter;
 import maple.expectation.monitoring.context.SystemContextProvider;
 import maple.expectation.monitoring.copilot.model.IncidentContext;
-import maple.expectation.monitoring.security.PiiMaskingFilter;
 import maple.expectation.monitoring.throttle.AlertThrottler;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
