+++ module-app/src/test-legacy/java/maple/expectation/global/error/GlobalExceptionHandlerTest.java
@@ -15,9 +15,9 @@ import java.util.concurrent.CompletionException;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeoutException;
 import maple.expectation.controller.dto.admin.AddAdminRequest;
-import maple.expectation.global.error.exception.ApiTimeoutException;
-import maple.expectation.global.error.exception.CharacterNotFoundException;
-import maple.expectation.global.error.GlobalExceptionHandler;
+import maple.expectation.error.exception.ApiTimeoutException;
+import maple.expectation.error.exception.CharacterNotFoundException;
+import maple.expectation.error.GlobalExceptionHandler;
 import maple.expectation.global.security.AuthenticatedUser;
 import maple.expectation.service.v2.auth.AdminService;
 import maple.expectation.service.v2.facade.GameCharacterFacade;
